books/pyproject.toml
[project]
name = "books"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strawberry-graphql[cli,debug-serer,debug-server]>=0.248.1",
]
